+++ src/services/Terms.php
@@ -5,6 +5,7 @@ namespace codemonauts\glossary\services;
 use codemonauts\glossary\elements\Glossary;
 use codemonauts\glossary\elements\Term;
 use Craft;
+use craft\helpers\ArrayHelper;
 use craft\helpers\Html;
 use Exception;
 use Twig\Error\SyntaxError;
@@ -13,7 +14,8 @@ use function Symfony\Component\String\s;
 
 class Terms extends Component
 {
-    protected $renderedTerms = '';
+    protected string $renderedTerms = '';
+    protected array $usedTerms = [];
 
     /**
      * Returns all terms to search for.
@@ -28,12 +30,12 @@ class Terms extends Component
             $term->term,
         ];
 
-        if ($term->synonyms !== '') {
+        if (!empty($term->synonyms)) {
             $synonyms = explode(',', $term->synonyms);
             $terms = array_merge($terms, $synonyms);
         }
 
-        return $terms;
+        return ArrayHelper::filterEmptyStringsFromArray($terms);
     }
 
     /**
@@ -50,10 +52,9 @@ class Terms extends Component
         $originalText = $text;
 
         try {
-            $termTemplate = $glossary->termTemplate !== '' ? $glossary->termTemplate : '<span>{{ text }}</span>';
+            $termTemplate = !empty($glossary->termTemplate) ? $glossary->termTemplate : '<span>{{ text }}</span>';
             $replacements = [];
             $terms = Term::find()->glossary($glossary)->all();
-            $usedTerms = [];
 
             foreach ($terms as $term) {
                 $template = Html::modifyTagAttributes($termTemplate, [
@@ -74,7 +75,7 @@ class Terms extends Component
                     if (!$term->caseSensitive) {
                         $pattern .= 'i';
                     }
-                    $text = s($text)->replaceMatches($pattern, function ($matches) use ($term, $template, &$replacements, &$index, $view, &$usedTerms, $glossary) {
+                    $text = s($text)->replaceMatches($pattern, function ($matches) use ($term, $template, &$replacements, &$index, $view, $glossary) {
                         try {
                             $replacement = trim($view->renderString($template, [
                                 'term' => $term,
@@ -99,7 +100,7 @@ class Terms extends Component
                         $variables['term'] = $term;
 
                         try {
-                            $usedTerms[$term->id] = $view->renderTemplate($glossary->tooltipTemplate, $variables, 'site');
+                            $this->usedTerms[$term->id] = $view->renderTemplate($glossary->tooltipTemplate, $variables, 'site');
                         } catch (SyntaxError $e) {
                             Craft::error($e->getMessage(), 'glossary');
                         }
@@ -114,7 +115,7 @@ class Terms extends Component
             }
 
             $renderedTerms = '';
-            foreach ($usedTerms as $id => $usedTerm) {
+            foreach ($this->usedTerms as $id => $usedTerm) {
                 $renderedTerms .= Html::tag('div', $usedTerm, [
                     'id' => 'term-' . $id,
                 ]);